move MergeList from pkg instanceset to controllerutil
